Allow message to terminate in Processor
Originally Processor was designed to accept one input and
produce one output to keep it simple and clean. In real world
scenario, say a Processor needs to do de-duplication, it should
not produce an output and leave the filtering to next component
along the pipeline. Message.terminates() added to address this.
Calling Message.terminates() will make Processor aware there will
be no output for this message. This change is an interface change,
code based on original Processor needs to be updated.

Add .get() and .update() for Message
Add .terminates() for Message
Change .process() to take Message argument instead of Message.dct
Skip output if Message.terminated is True
Add pylint